% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{basic_logistic_reg}
\alias{basic_logistic_reg}
\title{Logistic Regression}
\usage{
basic_logistic_reg(x, y)
}
\arguments{
\item{x}{the design matrix, with predictors as columns.}

\item{y}{the response vector.}
}
\value{
vector of coefficients.
}
\description{
basic_logistic_reg function that performs logistic regression, implemented here for baseline unit testing.
}
