% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_logistic_reg.R
\name{logistic_reg}
\alias{logistic_reg}
\title{Logistic Regression}
\usage{
logistic_reg(
  x,
  y,
  start = rep(0, ncol(x)),
  eps_f = 1e-08,
  eps_g = 1e-05,
  maxit = 300
)
}
\arguments{
\item{x}{the design matrix, with predictors as columns.}

\item{y}{the response vector.}

\item{start}{The initial guess of the coefficient vector.}

\item{eps_f}{Iteration stops if \eqn{|f-f'|/|f|<\epsilon_f}{|f-f'|/|f|<eps_f},
where \eqn{f} and \eqn{f'} are the current and previous value
of the objective function (negative log likelihood) respectively.}

\item{eps_g}{Iteration stops if
\eqn{||g|| < \epsilon_g * \max(1, ||\beta||)}{||g|| < eps_g * max(1, ||beta||)},
where \eqn{\beta}{beta} is the current coefficient vector and
\eqn{g} is the gradient.}

\item{maxit}{Maximum number of iterations.}
}
\value{
a list with the following components:
\item{coefficients}{Coefficient vector}
\item{fitted.values}{The fitted probability values}
\item{linear.predictors}{The fitted values of the linear part, i.e.,
                         \eqn{X\hat{\beta}}{X * beta_hat}}
\item{loglikelihood}{The maximized log likelihood}
\item{converged}{Whether the optimization algorithm has converged}
}
\description{
logistic_reg function that performs logistic regression, implemented here for baseline unit testing.
}
