% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRpadding-package.R
\docType{package}
\name{LRpadding-package}
\alias{LRpadding}
\alias{LRpadding-package}
\title{LRpadding: Padded Logistic Regression}
\description{
Part of a larger project developed here for testing purposes, implements here a padded logistic regression.
}
\author{
\strong{Maintainer}: Christopher Gill \email{chrgill@gmail.com}

}
\keyword{internal}
