% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padded_logistic_reg.R
\name{padded_logistic_reg}
\alias{padded_logistic_reg}
\title{Padded Logistic Regression}
\usage{
padded_logistic_reg(x, y, padding = 0)
}
\arguments{
\item{x}{the design matrix, with predictors as columns.}

\item{y}{the response vector.}

\item{padding}{an integer defining the padding level to be used (default is zero).}
}
\value{
vector of coefficients.
}
\description{
padded_logistic_reg function that performs padded logistic regression.  
This is equivalent to using fastLR in the RcppNumerical package having
 padded predictors and the response with 0's.
}
