% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_logreg_coeffs}
\alias{update_logreg_coeffs}
\title{LR Update Function}
\usage{
update_logreg_coeffs(
  model,
  responsemat,
  coeffmat,
  Xb,
  fitted,
  cols_to_update,
  padding_zeros = 0L
)
}
\arguments{
\item{model}{- model matrix.}

\item{responsemat}{matrix of response vectors in columns.}

\item{coeffmat}{initial coefficient matrix.}

\item{Xb}{matrix of linear predictors, one column for each response.}

\item{fitted}{matrix of fitted probabilities.}

\item{cols_to_update}{vector of zeros and ones (doubles) indicating which
columns to update.}

\item{padding_zeros}{integer number of padding zeros to add to the response
and predictors to update.}
}
\value{
list of coefficient matrix, predictor matrix, fitted_values matrix,
vector of old loglikelihoods (from inputs) and vector on new log likelihoods
based on the returned coefficients.
}
\description{
Runs Logistic Regression on each response with the model and updates the
coefficients, predictors, and fitted values when the log-likelihood has
been improved.  Updates values in place.
}
